Added tests for Employee Type
