+++ EmpApp/EmpApp/DataModel/Employee.swift
@@ -33,22 +33,3 @@ struct EmployeeElement: Codable {
         case employeeType = "employee_type"
     }
 }
-
-// MARK: - EmployeeType
-enum EmployeeType: String, Codable {
-    case contractor = "CONTRACTOR"
-    case fullTime = "FULL_TIME"
-    case partTime = "PART_TIME"
-    
-    var shortDescription: String {
-        switch self {
-        case .contractor:
-            return "C"
-        case .fullTime:
-            return "FTE"
-        case .partTime:
-            return "PT"
-        }
-    }
-}
-
